♻️ refactor(controller): regroup by domain behind a barrel
Controller mixed widget-kind folders (Button/Input/Drawer) with domain
folders (Volume/SpeedSelector), so editing one feature meant touching two
unrelated directories — the volume icon lived under Button/ while the rest
of volume lived under Volume/.

Regroup into Transport/Progress/Volume/Speed/PlayList and let
Controller/index.tsx re-export the slot components. The composition root
previously imported five deep paths into Controller internals, which tied
the public compound slots (AudioPlayer.Progress and friends) to internal
folder names; it now imports a single path.

Behavior unchanged: 451 tests pass, typecheck, lint and build are clean,
and the bundle exports the same 28 symbols.
