--- package/src/components/AudioPlayer/Interface/Controller/Button/index.ts
@@ -1,6 +1,0 @@
-export * from "./TransportControls";
-export * from "./TransportControls/PlayBtn";
-export * from "./TransportControls/PrevBtn";
-export * from "./TransportControls/NextBtn";
-export * from "./RepeatTypeBtn";
-export * from "./VolumeIcon";
