--- package/src/components/AudioPlayer/Interface/Controller/Drawer/index.ts
@@ -1,1 +1,0 @@
-export * from "./SortablePlayList";
