--- package/src/components/AudioPlayer/Interface/Controller/Input/index.ts
@@ -1,1 +1,0 @@
-export * from "./Progress";
