+++ package/src/components/AudioPlayer/Interface/Controller/PlayList/PlayListIcon.tsx
@@ -1,7 +1,7 @@
 import { useResourceContext } from "@/components/AudioPlayer/Context/hooks/useResourceContext";
 import { FC } from "react";
 import { MdPlaylistPlay } from "@/components/icons";
-import { Icon } from "../../Icon";
+import { Icon } from "../Icon";
 
 export const PlayListIcon: FC = () => {
   const { customIcons } = useResourceContext();
