+++ package/src/components/AudioPlayer/Interface/Controller/Progress/BarProgress.tsx
