+++ package/src/components/AudioPlayer/Interface/Controller/Progress/WaveformProgress.tsx
