+++ package/src/components/AudioPlayer/Interface/Controller/Progress/__tests__/BarProgress.a11y.test.tsx
