+++ package/src/components/AudioPlayer/Interface/Controller/Progress/__tests__/BarProgress.tooltip.test.tsx
