+++ package/src/components/AudioPlayer/Interface/Controller/Progress/__tests__/ProgressTooltip.test.tsx
