+++ package/src/components/AudioPlayer/Interface/Controller/Progress/__tests__/WaveformProgress.test.tsx
