+++ package/src/components/AudioPlayer/Interface/Controller/Progress/__tests__/useElementWidth.test.tsx
