+++ package/src/components/AudioPlayer/Interface/Controller/Progress/__tests__/useWavesurfer.test.tsx
