+++ package/src/components/AudioPlayer/Interface/Controller/Progress/hooks/useElementWidth.ts
