+++ package/src/components/AudioPlayer/Interface/Controller/Progress/hooks/useWaveformMode.ts
