+++ package/src/components/AudioPlayer/Interface/Controller/Progress/hooks/useWavesurfer.ts
