+++ package/src/components/AudioPlayer/Interface/Controller/Speed/__tests__/SpeedSelector.a11y.test.tsx
