+++ package/src/components/AudioPlayer/Interface/Controller/Transport/index.tsx
@@ -1,10 +1,15 @@
 import { FC } from "react";
 import Grid, { GridItemLayoutProps } from "@/components/Grid";
 import { useUIContext } from "@/components/AudioPlayer/Context/hooks/useUIContext";
-import { useResolvedGridArea } from "../../../hooks/useResolvedGridArea";
+import { useResolvedGridArea } from "../../hooks/useResolvedGridArea";
 import { PlayBtn } from "./PlayBtn";
 import { PrevBtn } from "./PrevBtn";
 import { NextBtn } from "./NextBtn";
+
+export * from "./PlayBtn";
+export * from "./PrevBtn";
+export * from "./NextBtn";
+export * from "./RepeatTypeBtn";
 
 export type TransportControlsProps = GridItemLayoutProps;
