+++ package/src/components/AudioPlayer/Interface/Controller/Volume/index.tsx
@@ -2,7 +2,7 @@ import { FC, useCallback, useRef } from "react";
 import { VolumeSlider } from "./Content";
 import Dropdown from "@/components/Dropdown";
 import Grid, { GridItemLayoutProps } from "@/components/Grid";
-import { VolumeIcon } from "../Button";
+import { VolumeIcon } from "./VolumeIcon";
 import { VolumeSliderPlacement } from "@/components/AudioPlayer/Context/StateContext";
 import { useUIContext } from "@/components/AudioPlayer/Context/hooks/useUIContext";
 import { usePlaybackContext } from "@/components/AudioPlayer/Context/hooks/usePlaybackContext";
@@ -10,6 +10,8 @@ import { useNonNullableContext } from "@/hooks/useNonNullableContext";
 import { audioPlayerDispatchContext } from "@/components/AudioPlayer/Context/dispatchContext";
 import { useResolvedGridArea } from "../../hooks/useResolvedGridArea";
 import { useResolvedDropdownProps } from "../../hooks/useResolvedDropdownProps";
+
+export * from "./VolumeIcon";
 
 export interface VolumeProps extends GridItemLayoutProps {
   triggerType?: "click" | "hover";
